EDGEAI-1213: Move fused draw_masks from Decoder to ImageProcessor
ImageProcessor owns the GPU context — it should orchestrate rendering.
Decoder decodes. The fused decode+render path was on the wrong object.

API changes (all layers):
- Removed: Decoder.draw_masks() (Python)
- Removed: hal_decoder_draw_masks / hal_decoder_draw_masks_tracked (C)
- Added: ImageProcessor.draw_masks_fused(decoder, model_output, dst, ...) (Python)
- Added: hal_image_processor_draw_masks_fused[_tracked](...) (C)

The fused path now lives on ImageProcessor and borrows the Decoder to
access its decode_*_proto methods. The existing pre-decoded
ImageProcessor.draw_masks(dst, boxes, scores, classes, seg, ...) is
unchanged.

Breaking Python and C API change (pre-1.0):
  # Before:
  decoder.draw_masks(outputs, processor, dst)
  # After:
  processor.draw_masks_fused(decoder, outputs, dst)
