+++ ARCHITECTURE.md
@@ -559,7 +559,7 @@ The HAL provides two workflows for consuming these masks:
 
 | Workflow | Python | Rust | C | CPU | OpenGL | G2D |
 |----------|--------|------|---|:---:|:------:|:---:|
-| **Draw** — fused overlay onto image | `decoder.draw_masks()` | `draw_masks_proto()` | `hal_decoder_draw_masks()` | Yes | Yes | No |
+| **Draw** — fused overlay onto image | `processor.draw_masks_fused()` | `draw_masks_proto()` | `hal_image_processor_draw_masks_fused()` | Yes | Yes | No |
 | **Draw pre-decoded** — draw already-decoded masks | `processor.draw_masks()` | `draw_masks()` | `hal_image_processor_draw_masks()` | Yes | Yes | No |
 
 > **G2D limitation:** The NXP G2D hardware accelerator does not support mask
@@ -572,7 +572,7 @@ The HAL provides two workflows for consuming these masks:
 
 | Use case | Recommended API | Why |
 |----------|----------------|-----|
-| Overlay colored masks onto a display frame | `decoder.draw_masks()` | Fused path — masks never leave Rust/GPU, lowest latency |
+| Overlay colored masks onto a display frame | `processor.draw_masks_fused()` | Fused path — masks never leave Rust/GPU, lowest latency |
 | Draw masks you already have (e.g. from a previous `decode()` call) | `processor.draw_masks()` | Accepts pre-decoded `(H, W, C)` mask arrays |
 
 **Format requirements:**
